binary.Write不给传int

+++ app.go
@@ -75,9 +75,9 @@ func augur(fromQQ int64) int32 {
 	//将被占卜物代入天命术式
 	must(binary.Write(hash, binary.BigEndian, fromQQ)) // never returns a err
 	y, m, d := time.Now().Date()
-	must(binary.Write(hash, binary.BigEndian, y))
-	must(binary.Write(hash, binary.BigEndian, m))
-	must(binary.Write(hash, binary.BigEndian, d))
+	must(binary.Write(hash, binary.BigEndian, int64(y)))
+	must(binary.Write(hash, binary.BigEndian, int64(m)))
+	must(binary.Write(hash, binary.BigEndian, int64(d)))
 
 	//用天命开始占卜
 	destiny := hash.Sum(nil) //获取destiny 天命
